# docs/swagger.yaml
definitions:
...
  model.MySQLTime:
    properties:
      time.Time:
        type: string
    type: object
...
  model.Wishlist:
    properties:
      created_at:
        $ref: '#/definitions/model.MySQLTime'
      game_id:
        type: integer
      id:
        type: integer
      updated_at:
        $ref: '#/definitions/model.MySQLTime'
      user_id:
        type: integer
    type: object
  model.WishlistWithGameTitle:
    properties:
      created_at:
        $ref: '#/definitions/model.MySQLTime'
      game_title:
        type: string
      id:
        type: integer
      updated_at:
        $ref: '#/definitions/model.MySQLTime'
      user_id:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /wishlist:
    get:
      description: Get a list of all wishlist items for the authenticated user
      responses:
        "200":
          description: List of wishlist items
          schema:
            items:
              $ref: '#/definitions/model.WishlistWithGameTitle'
            type: array
        "401":
          description: Unauthorized" (when the JWT token is missing or invalid)
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error" (when there is a problem with the database)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all wishlist items
    post:
      description: Add a new wishlist item for the authenticated user
      parameters:
      - description: Wishlist object that needs to be added
        in: body
        name: wishlist
        required: true
        schema:
          $ref: '#/definitions/model.Wishlist'
      responses:
        "200":
          description: Wishlist item added
          schema:
            additionalProperties: true
            type: object
        "400":
          description: The Game is already exists in your list" (when the game is
            already present in the user's wishlist)
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized" (when the JWT token is missing or invalid)
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error" (when there is a problem with the database)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Add new wishlist
  /wishlist/{id}:
    delete:
      description: Delete a wishlist item by its ID
      parameters:
      - description: Wishlist item ID to delete
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Wishlist item successfully deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid wishlist item ID" (when the wishlist item ID in the
            URL path is not a valid integer)
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized" (when the JWT token is missing or invalid)
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Wishlist item not found" (when the specified wishlist item
            ID does not exist in the database)
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error" (when there is a problem with the database)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete wishlist item by ID
swagger: "2.0"